# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LitChat API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterBody {
  /**
   * @minLength 3
   * @maxLength 20
   */
  username: string;
  /** @minLength 6 */
  password: string;
  avatar?: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface UpdateAvatarBody {
  avatar: string;
}

export interface User {
  id: number;
  username: string;
  avatar: string;
  status: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type MessageReactions = { [key: string]: string[] };

export interface Message {
  id: number;
  room: string;
  username: string;
  avatar: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  audioUrl?: string | null;
  messageType: string;
  /** @nullable */
  replyToId?: number | null;
  /** @nullable */
  editedAt?: string | null;
  isDeleted: boolean;
  reactions?: MessageReactions;
  createdAt: string;
}

export interface Friendship {
  id: number;
  requester: string;
  recipient: string;
  status: string;
  createdAt: string;
}

export type GetRoomMessagesParams = {
  limit?: number;
};