# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LitChat API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: users
    description: User operations
  - name: messages
    description: Message operations
  - name: friends
    description: Friend operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{username}/avatar:
    patch:
      operationId: updateAvatar
      tags: [users]
      summary: Update user avatar
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvatarBody"
      responses:
        "200":
          description: Avatar updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /messages/{room}:
    get:
      operationId: getRoomMessages
      tags: [messages]
      summary: Get messages for a room
      parameters:
        - name: room
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

  /friends:
    get:
      operationId: getFriends
      tags: [friends]
      summary: Get my friends list
      responses:
        "200":
          description: Friends list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Friendship"

  /friends/requests:
    get:
      operationId: getFriendRequests
      tags: [friends]
      summary: Get incoming friend requests
      responses:
        "200":
          description: Friend requests list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Friendship"

  /friends/request/{username}:
    post:
      operationId: sendFriendRequest
      tags: [friends]
      summary: Send a friend request
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Friendship"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already friends or pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /friends/accept/{requestId}:
    post:
      operationId: acceptFriendRequest
      tags: [friends]
      summary: Accept a friend request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Friendship"
        "404":
          description: Request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /friends/{requestId}:
    delete:
      operationId: removeFriend
      tags: [friends]
      summary: Remove friend or decline request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    RegisterBody:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
        password:
          type: string
          minLength: 6
        avatar:
          type: string
      required:
        - username
        - password

    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    UpdateAvatarBody:
      type: object
      properties:
        avatar:
          type: string
      required:
        - avatar

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token

    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        avatar:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - avatar
        - status
        - createdAt

    Message:
      type: object
      properties:
        id:
          type: integer
        room:
          type: string
        username:
          type: string
        avatar:
          type: string
        content:
          type: ["string", "null"]
        audioUrl:
          type: ["string", "null"]
        messageType:
          type: string
        replyToId:
          type: ["integer", "null"]
        editedAt:
          type: ["string", "null"]
          format: date-time
        isDeleted:
          type: boolean
        reactions:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - room
        - username
        - avatar
        - messageType
        - isDeleted
        - createdAt

    Friendship:
      type: object
      properties:
        id:
          type: integer
        requester:
          type: string
        recipient:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - requester
        - recipient
        - status
        - createdAt