# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LitChat API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const registerBodyUsernameMin = 3;
export const registerBodyUsernameMax = 20;

export const registerBodyPasswordMin = 6;

export const RegisterBody = zod.object({
  username: zod
    .string()
    .min(registerBodyUsernameMin)
    .max(registerBodyUsernameMax),
  password: zod.string().min(registerBodyPasswordMin),
  avatar: zod.string().optional(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    username: zod.string(),
    avatar: zod.string(),
    status: zod.string(),
    createdAt: zod.coerce.date(),
  }),
  token: zod.string(),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  avatar: zod.string(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update user avatar
 */
export const UpdateAvatarParams = zod.object({
  username: zod.coerce.string(),
});

export const UpdateAvatarBody = zod.object({
  avatar: zod.string(),
});

export const UpdateAvatarResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  avatar: zod.string(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get messages for a room
 */
export const GetRoomMessagesParams = zod.object({
  room: zod.coerce.string(),
});

export const getRoomMessagesQueryLimitDefault = 50;

export const GetRoomMessagesQueryParams = zod.object({
  limit: zod.coerce.number().default(getRoomMessagesQueryLimitDefault),
});

export const GetRoomMessagesResponseItem = zod.object({
  id: zod.number(),
  room: zod.string(),
  username: zod.string(),
  avatar: zod.string(),
  content: zod.string().nullish(),
  audioUrl: zod.string().nullish(),
  messageType: zod.string(),
  replyToId: zod.number().nullish(),
  editedAt: zod.coerce.date().nullish(),
  isDeleted: zod.boolean(),
  reactions: zod.record(zod.string(), zod.array(zod.string())).optional(),
  createdAt: zod.coerce.date(),
});
export const GetRoomMessagesResponse = zod.array(GetRoomMessagesResponseItem);

/**
 * @summary Get my friends list
 */
export const GetFriendsResponseItem = zod.object({
  id: zod.number(),
  requester: zod.string(),
  recipient: zod.string(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
});
export const GetFriendsResponse = zod.array(GetFriendsResponseItem);

/**
 * @summary Get incoming friend requests
 */
export const GetFriendRequestsResponseItem = zod.object({
  id: zod.number(),
  requester: zod.string(),
  recipient: zod.string(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
});
export const GetFriendRequestsResponse = zod.array(
  GetFriendRequestsResponseItem,
);

/**
 * @summary Send a friend request
 */
export const SendFriendRequestParams = zod.object({
  username: zod.coerce.string(),
});

/**
 * @summary Accept a friend request
 */
export const AcceptFriendRequestParams = zod.object({
  requestId: zod.coerce.number(),
});

export const AcceptFriendRequestResponse = zod.object({
  id: zod.number(),
  requester: zod.string(),
  recipient: zod.string(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Remove friend or decline request
 */
export const RemoveFriendParams = zod.object({
  requestId: zod.coerce.number(),
});

export const RemoveFriendResponse = zod.object({
  success: zod.boolean(),
});