# lib/api-zod/src/generated/types/friendship.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LitChat API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Friendship {
  id: number;
  requester: string;
  recipient: string;
  status: string;
  createdAt: Date;
}