# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LitChat API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MessageReactions } from "./messageReactions";

export interface Message {
  id: number;
  room: string;
  username: string;
  avatar: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  audioUrl?: string | null;
  messageType: string;
  /** @nullable */
  replyToId?: number | null;
  /** @nullable */
  editedAt?: Date | null;
  isDeleted: boolean;
  reactions?: MessageReactions;
  createdAt: Date;
}